Not recompile if file osrm exists

+++ install.sh
@@ -5,20 +5,29 @@
 export PROFILE_LUA=$(find /opt/osrm/profile/ -type f -name '*.lua')
 export DATA_OSM_PBF=$(find /opt/osrm/data/ -type f -name '*.pbf')
 
-echo $PROFILE_LUA
-echo $DATA_OSM_PBF
+export DATA_OSM_OSRM=$(find /opt/osrm/data/ -type f -name '*.osrm')
 
-cd /opt/osrm/osrm-backend/build
+if [ ! -f $DATA_OSM_OSRM ]
+	then
+	echo $PROFILE_LUA
+	echo $DATA_OSM_PBF
 
-ln -s $PROFILE_LUA profile.lua
-ln -s ../profiles/lib/
+	cd /opt/osrm/osrm-backend/build
 
-./osrm-extract -p profile.lua $DATA_OSM_PBF
+	ln -s $PROFILE_LUA profile.lua
+	ln -s ../profiles/lib/
+
+	./osrm-extract -p profile.lua $DATA_OSM_PBF
+
+	export DATA_OSM_OSRM=$(find /opt/osrm/data/ -type f -name '*.osrm')
+
+	echo $DATA_OSM_OSRM
+
+	./osrm-contract $DATA_OSM_OSRM
+fi
 
 export DATA_OSM_OSRM=$(find /opt/osrm/data/ -type f -name '*.osrm')
 
 echo $DATA_OSM_OSRM
-
-./osrm-contract $DATA_OSM_OSRM
 
 ./osrm-routed $DATA_OSM_OSRM --max-table-size=2000
